% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgl.ball.R
\name{rgl.ball}
\alias{rgl.ball}
\title{Plot high-resolution sphere / globe}
\usage{
rgl.ball(x = 0, y = 0, z = 0, radius = 1, depth = 4, png = NULL, ...)
}
\arguments{
\item{x}{x-coordinate of the center of the sphere}

\item{y}{y-coordinate of the center of the sphere}

\item{z}{z-coordinate of the center of the sphere}

\item{radius}{radius of the sphere}

\item{depth}{integer (1...7) specifying the number of rectangles (=6*4^depth)}

\item{png}{optional character string specifying the file name of a png-image to be rendered on the sphere. This file must contain the map to be displayed in an equirectangular projection (also known as equidistant cylindrical projection).}

\item{...}{additional parameter to refine the material properties (see \code{\link[rgl]{rgl.material}}).}
}
\description{
Draws a sphere in custom resolution with custom surface image.
}
\examples{
# Show Earth with core
rgl.new(width=0.5, aspect=1.0, col='black', xlim=c(-1,1), ylim=c(-1,1), zlim=c(-1,1))
rgl.ball(0, 0, 0, 1, png=system.file('earth.png', package='rglplus'), emission='grey', alpha=0.5)
rgl.ball(0, 0, 0, 0.6, col='red')
rgl.orthoview('yz', zoom=0.5)

}
\author{
Danail Obreschkow (thanks to input from Aaron Robotham's \code{sphereplot} package)
}
