% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgl.camera.R
\name{rgl.camera}
\alias{rgl.camera}
\title{Place observer}
\usage{
rgl.camera(position = 10, direction = NULL, angle = 0, fov = 0)
}
\arguments{
\item{position}{either a single number, a 3-vector or NULL. A single number is interpreted as the distance between the observer and the center of the scene. The observer is placed at distance along the line-of-sight of the current/specified viewing direction. A 3-vector is interpreted as the actual position of the observer in the coordinates of the 3D plot. If \code{NULL}, the position is adjusted automatically.}

\item{direction}{optional 3-vector specifying the direction in which the observer is looking. The norm of the vector is irrelevant, but has to be larger than zero.}

\item{angle}{(only used if \code{direction} is specified) position angle in degrees, defining the rotation of the camera around the line-of-sight. This angle is defined as the angle formed between the horizontal line on the screen and the projected z-axis. Only the z-axis gets very close to the line-of-sight, the angle refers to the projected y-axis.}

\item{fov}{(only used if \code{direction} is specified) field of view in degrees, as used in \code{\link[rgl]{rgl.viewpoint}}. This is roughly the field-of-view seen along the shortest axis of the window.}
}
\description{
Set the position, orientation and field-of-view of the observer/camera.
}
\examples{
rgl.test.scene()
rgl.camera(position=c(1.4,1.4,1.4), direction=c(-1,-1,-1), fov=60, angle=0)

}
\author{
Danail Obreschkow
}
