% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgl.close.all.R
\name{rgl.close.all}
\alias{rgl.close.all}
\title{Close all open rgl windows}
\usage{
rgl.close.all()
}
\value{
None
}
\description{
Checks if any rgl windows are currently open and, if so, closes them.
}
\author{
Danail Obreschkow
}
