% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgl.draw.R
\name{rgl.draw}
\alias{rgl.draw}
\title{Continue drawing on screen}
\usage{
rgl.draw()
}
\description{
Updates screen display after this was stopped using the function \code{\link{rgl.hold}}.
}
\seealso{
\code{\link{rgl.draw}}
}
\author{
Danail Obreschkow
}
