% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgl.hold.R
\name{rgl.hold}
\alias{rgl.hold}
\title{Stop drawing on screen}
\usage{
rgl.hold()
}
\value{
None
}
\description{
Prevents the following rgl functions from drawing on the screen, until the function \code{\link{rgl.draw}} is called. This is used to accelerate complex drawings. This routine is identical to calling \code{par3d(skipRedraw=TRUE)}.
}
\seealso{
\code{\link{rgl.draw}}
}
\author{
Danail Obreschkow
}
