% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgl.makemovie.R
\name{rgl.makemovie}
\alias{rgl.makemovie}
\title{Produce a movie from and 3d rgl scene}
\usage{
rgl.makemovie(
  frame = NULL,
  path = NULL,
  tmin = NULL,
  tmax = NULL,
  nframes = 60,
  fps = 60,
  output.path,
  output.filename,
  keep.frames = F,
  quiet = T,
  separator = "/",
  ffmpeg.cmd = "ffmpeg",
  ffmpeg.opt = "-vcodec libx264 -crf 18 -pix_fmt yuv420p"
)
}
\arguments{
\item{frame}{optional function that plots or updates the scene at a given time. This function must have exactly one  argument, which specifies the time of the frame.}

\item{path}{optional list that specifies the motion of the camera at some discrete times. The list contains the following elements (for more details see \code{\link{rgl.camera}}):\cr\cr
\code{time} = optional n-vector of strictly monotonically increasing discrete times; if not given, the other arguments (potion, direction, up, fov) are assumed equally spaced in time.\cr
\code{position} = n-by-3 matrix of n discrete camera positions at the specified times; can also be a 3-vector specifying a constant position.\cr
\code{direction} = n-by-3 matrix of n discrete camera directions at the specified times; can also be a 3-vector specifying a constant direction. If not given, the tangent vector of the path is used as direction.\cr
\code{up} = n-by-3 matrix of n discrete up-directions of the camera; can also be a 3-vector specifying a constant up-direction. If not given, the (outward facing) normal vector of the path is used as up-direction.\cr
\code{fov} = n-vector of field of view (in degrees) at the specified times; an also be a single number specifying a constant field of view.\cr\cr}

\item{tmin}{time of first frame in the movie. Should normally not be smaller than the first time in path$time. If not specified, the first value in path$time is adopted.}

\item{tmax}{time of last frame in the movie. Should normally not be larger than the last time in path$time. If not specified, the last value in path$time is adopted.}

\item{nframes}{number of frames in the movie. The time variable is sampled evenly between \code{tmin} and \code{tmax}.}

\item{fps}{number of frames per second}

\item{output.path}{character specifying the directory, where the movie and temporary frames are saved}

\item{output.filename}{movie filename without path. This filename should end on the extension '.mp4'.}

\item{keep.frames}{logical flag specifying whether the temporary directory with the individual frame files should be kept}

\item{quiet}{logical flag; if true, all console outputs produced by 'ffmpeg' are suppressed}

\item{separator}{filename separate of the system ('/' for Mac, Linux, Unix; '\' for Windows)}

\item{ffmpeg.cmd}{command used to call ffmpeg form a terminal. Normally, this is just 'ffmpeg'.}

\item{ffmpeg.opt}{compression and formatting options used with ffmpeg}
}
\description{
Generates an MP4-movie of a 3d rgl scene with time-dependent objects and/or a camera path. The routine has been developed and tested for MacOS and it requires on a working installation of ffmpeg.
}
\details{
Note that the frame width and height should be divisible by 2 for mp4 video compression to work.\cr
To accelerate the movie generation, it is possible to suppress the screen update by calling \code{\link{rgl.hold}} before calling \code{rgl.makemovie}.
}
\examples{

## Example: Movie of spaceship flying around the Earth chasing a UFO

# Produce basic scene
rgl.new(width=720, aspect=4/3, col='black', xlim=c(-1,1), ylim=c(-1,1), zlim=c(-1,1))
rgl::clear3d(type = "lights")
rgl::rgl.light(80,30,viewpoint.rel = FALSE)
rgl.ball(0, 0, 0, 1, png=system.file('earth.png', package='rglplus'), depth=7, emission='#444466')

# Make function to update frames
ptlast = NULL
frame = function(t) {
  rgl.hold()
  pt = c(1.01*cos(t+0.2),1.01*sin(t+0.2),sin(5*t)*0.02)
  if (!is.null(ptlast)) {
    rgl::pop3d()
    rgl::lines3d(rbind(ptlast,pt), col='red', lwd=3)
  }
  rgl.ball(pt[1], pt[2], pt[3], 0.003, col='orange', emission='#885500')
  ptlast <<- pt
  rgl.draw()
}

# Make path
alpha = seq(0,2*pi,length=37)
path = list(position=cbind(1.1*cos(alpha),1.1*sin(alpha),0),
            direction=cbind(-1.5*sin(alpha)-cos(alpha),1.5*cos(alpha)-sin(alpha),0),
            up=cbind(cos(alpha),sin(alpha),sin(alpha)/2),
            fov=30)

# Produce movie
\dontrun{
rgl.makemovie(frame=frame, path=path, tmin=0, tmax=2*pi, output.path='~/testmovie',
              output.filename = 'movie.mp4', ffmpeg.cmd = 'ffmpeg', nframes=600)
}

}
\author{
Danail Obreschkow
}
