% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgl.test.scene.R
\name{rgl.test.scene}
\alias{rgl.test.scene}
\title{Plot 3D test image}
\usage{
rgl.test.scene(center = c(0, 0, 0))
}
\arguments{
\item{center}{3-vector specifying the centre of the 3D plot.}
}
\value{
None
}
\description{
Draws a c3D test image with three Cartesian axes, a sphere of radius 0.5 and three light sources with RGB colors.
}
\examples{
rgl.test.scene()

}
\author{
Danail Obreschkow
}
