% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rglplus-package.R
\docType{package}
\name{rglplus}
\alias{rglplus}
\title{Extension of the 'rgl' 3D Visualization Package}
\description{
Provides 3D plotting routines that facilitate the use of the 'rgl' package and extend its functionality. For example, the routines allow the user to directly control of the camera position & orientation, as well as to generate 3D movies with a moving observer.
}
\author{
Danail Obreschkow
}
