% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgl.draw.R
\name{rgl.draw}
\alias{rgl.draw}
\title{Continue drawing on screen}
\usage{
rgl.draw()
}
\value{
None
}
\description{
Updates screen display after this was stopped using the function \code{\link{rgl.hold}}. This is identical to calling \code{par3d(skipRedraw=FALSE)}.
}
\seealso{
\code{\link{rgl.hold}}
}
\author{
Danail Obreschkow
}
