% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgl.test.scene.R
\name{rgl.test.scene}
\alias{rgl.test.scene}
\title{Plot 3D test image}
\usage{
rgl.test.scene(center = c(0, 0, 0), width = 0.5)
}
\arguments{
\item{center}{3-vector specifying the centre of the 3D plot.}

\item{width}{either an integer (>1) specifying the number of pixels in the horizontal direction, or a real value (>0 and <=1) specifying the fraction of the available pixels. If the selected aspect ratio causes the number of vertical pixels to exceed the available number, the width is reduced as much as necessary.}
}
\value{
None
}
\description{
Draws a c3D test image with three Cartesian axes, a sphere of radius 0.5 and three light sources with RGB colors.
}
\author{
Danail Obreschkow
}
