% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgl.test.scene.R
\name{rgl.test.scene}
\alias{rgl.test.scene}
\title{Plot 3D test image}
\usage{
rgl.test.scene()
}
\value{
None
}
\description{
Draws a cubic 3D test image with three Cartesian axes, a sphere and three light sources with RGB colors.
}
\examples{
\donttest{
rgl.test.scene()
}

}
\author{
Danail Obreschkow
}
